Fix TimeInterval <duration>/<end>
The simptoms are: When the given pattern ends on December (e.g.
P1Y/2017-12-06T18:30:00Z) the result is off by roughly a year.

The first bug was related to trying to calculate the amount of seconds
using a positive period instead of a negative one. In other words, the
start date for P1Y/2017-12-06T18:30:00Z is:

start_date =  2017-12-06T18:30:00Z + -P1Y

But it was:

start_date =  2017-12-06T18:30:00Z - P1Y

Although the arithmetic are the same on the surface, P1Y and -P1Y
compute the amount in seconds in base of the given point in time AND the
direction of the duration (i.e. the sign of the duration).

Solving this error surfaced another one: The total seconds for a year
were calculated based on the differential of years instead of the
differential of points in time, so it was disregarding month which makes
the duration of a year different.

There is still a pending issue: Fractions of years are dropped so
P1.5Y/2017-12-06T18:30:00Z will fail.

Signed-off-by: Arnau Siches <asiches@gmail.com>