Reduce the number of files that ship in the artifact
The installed artifact for this gem is rather large. Since the install
is really just an artifact buried deep on a user's system there's no
need to ship the full test suite and documentation. This reduces the
install size, especially for packages products that use this gem. It's a
pretty common pattern to only ship the lib dir to keep the install size
down.

Signed-off-by: Tim Smith <tsmith@chef.io>